chore: move probe validation to prometheus module
